+++ packages/apidash_core/lib/services/http_client_manager.dart
@@ -1,8 +1,15 @@
-import 'package:apidash_core/utils/http_request_utils.dart';
+import 'dart:io';
+import 'dart:collection';
 import 'package:flutter/foundation.dart';
 import 'package:http/http.dart' as http;
-import 'dart:collection';
-import '';
+import 'package:http/io_client.dart';
+
+http.Client createHttpClientWithNoSSL() {
+  var ioClient = HttpClient()
+    ..badCertificateCallback =
+        (X509Certificate cert, String host, int port) => true;
+  return IOClient(ioClient);
+}
 
 class HttpClientManager {
   static final HttpClientManager _instance = HttpClientManager._internal();
@@ -16,13 +23,15 @@ class HttpClientManager {
 
   HttpClientManager._internal();
 
-  http.Client createClient(String requestId,{bool noSSL = false}) {
-    final client = noSSL && !kIsWeb ? createHttpClientWithNoSSL() :http.Client();
+  http.Client createClient(
+    String requestId, {
+    bool noSSL = false,
+  }) {
+    final client =
+        (noSSL && !kIsWeb) ? createHttpClientWithNoSSL() : http.Client();
     _clients[requestId] = client;
     return client;
   }
-
- 
 
   void cancelRequest(String? requestId) {
     if (requestId != null && _clients.containsKey(requestId)) {
